// apps/api/src/common/cognito/application/enum/cognito.enum.ts
export enum CognitoMessage {
  INVALID_REFRESH_TOKEN_ERROR = 'Invalid refresh token',
  PASSWORD_VALIDATION_ERROR = 'Password must be at least 8 characters long and contain at least one uppercase letter, one lowercase letter, one number and one special character',
  EXPIRED_CODE_ERROR = 'Code has expired',
  CODE_MISMATCH_ERROR = 'Code does not match',
  NEW_PASSWORD_REQUIRED_ERROR = 'New password required',
  INVALID_PASSWORD_ERROR = 'Incorrect email or password',
  USER_NOT_CONFIRMED_ERROR = 'User is not confirmed',
  INVALID_CODE_ERROR = 'Incorrect email or code',
  USER_EXISTS_ERROR = 'User already exists',
  USER_REGISTERED_SUCCESS = 'User registered successfully',
  USER_AUTHENTICATED_SUCCESS = 'User authenticated successfully',
  USER_REGISTRATION_CONFIRMED_SUCCESS = 'User registration confirmed successfully',
  CONFIRMATION_CODE_RESENT_SUCCESS = 'Confirmation code resent successfully',
  PASSWORD_RESET_SUCCESS = 'Password reset successfully',
  PASSWORD_RESET_CONFIRMATION_SUCCESS = 'Password reset confirmation successfully',
  USER_NOT_FOUND_ERROR = 'User not found',
  USER_EXISTS = 'User exists',
  PASSWORD_CHANGED_SUCCESS = 'Password changed successfully',
  USER_DOES_NOT_HAVE_A_SUB_ATTRIBUTE = 'User does not have a sub attribute',
  PASSWORD_RESET_INITIATED_SUCCESS = 'Password reset initiated successfully',
  PASSWORD_RESET_CONFIRMED_SUCCESS = 'Password reset confirmed successfully',
  LIMIT_EXCEEDED_ERROR = 'Limit exceeded',
  INVALID_PARAMETERS_ERROR = 'Invalid parameters',
  USERNAME_EXISTS_ERROR = 'Username exists',
  GENERIC_ERROR = 'An error occurred',
  USER_REGISTRATION_SUCCESS = 'User registration successful',
  COGNITO_LOCAL_UNSUPPORTED = 'Cognito local is not supported',
  USER_ALREADY_CONFIRMED = 'User already confirmed',
  SESSION_REFRESH_SUCCESS = 'Session refreshed successfully',
}

export enum CognitoError {
  LIMIT_EXCEEDED_EXCEPTION = 'LimitExceededException',
  INVALID_PASSWORD_EXCEPTION = 'InvalidPasswordException',
  USERNAME_EXISTS_EXCEPTION = 'UsernameExistsException',
  USER_NOT_CONFIRMED_EXCEPTION = 'UserNotConfirmedException',
  NOT_AUTHORIZED_EXCEPTION = 'NotAuthorizedException',
  PASSWORD_RESET_REQUIRED_EXCEPTION = 'PasswordResetRequiredException',
  CODE_MISMATCH_EXCEPTION = 'CodeMismatchException',
  EXPIRED_CODE_EXCEPTION = 'ExpiredCodeException',
  USER_NOT_FOUND_EXCEPTION = 'UserNotFoundException',
  INVALID_PARAMETER_EXCEPTION = 'InvalidParameterException',
  COGNITO_LOCAL_UNSUPPORTED = 'Cognito Local unsupported feature',
}